# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.changePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  handler.createUserRequest:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  handler.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handler.loginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/handler.userResponse'
    type: object
  handler.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  handler.resetPasswordRequest:
    properties:
      password:
        type: string
    type: object
...
  handler.updateUserRequest:
    properties:
      role:
        type: string
    type: object
  handler.userDetailResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
  handler.userListResponse:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/handler.userDetailResponse'
        type: array
    type: object
  handler.userResponse:
    properties:
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
  resp.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  resp.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: SQL 审核、查询和工单管理平台 API
  title: SQLFlow API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，返回 JWT token
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.loginResponse'
              type: object
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      summary: 用户登录
      tags:
      - 认证
  /auth/me:
    get:
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.userDetailResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /auth/password:
    put:
      consumes:
      - application/json
      description: 修改当前用户密码
      parameters:
      - description: 修改密码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码修改成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "401":
          description: 旧密码错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 refresh_token 获取新的 access_token
      parameters:
      - description: 刷新Token请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.loginResponse'
              type: object
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "401":
          description: refresh_token 无效或已过期
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      summary: 刷新Token
      tags:
      - 认证
...
  /users:
    get:
      description: 管理员获取所有用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.userListResponse'
              type: object
        "500":
          description: 获取用户列表失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 管理员创建新用户
      parameters:
      - description: 创建用户请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.userDetailResponse'
              type: object
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "500":
          description: 创建用户失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      description: 管理员删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "403":
          description: 不能删除自己或最后一个管理员
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      description: 管理员获取指定用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.userDetailResponse'
              type: object
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 管理员更新指定用户的角色
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新用户请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.userDetailResponse'
              type: object
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "403":
          description: 不能编辑自己的角色
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户角色
      tags:
      - 用户管理
  /users/{id}/reset-password:
    put:
      consumes:
      - application/json
      description: 管理员重置指定用户密码
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 重置密码请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码重置成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 重置用户密码
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"